Exception handling

+++ src/main/java/ru/kotb/accounting_system/controller/OrganisationController.java
@@ -5,10 +5,12 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.kotb.accounting_system.entity.Organisation;
+import ru.kotb.accounting_system.exception_handling.NoSuchEntityException;
 import ru.kotb.accounting_system.service.OrganisationService;
 
 import java.util.List;
@@ -22,16 +24,22 @@ import java.util.List;
 public class OrganisationController {
 
     /**
-     * The OrganisationService object for working with the
+     * The OrganisationService bean for working with the
      * organisations.
      */
     private final OrganisationService organisationService;
 
+    /**
+     * Constructs the controller and links it with the service bean.
+     *
+     * @param organisationService the organisation service bean
+     */
     @Autowired
     public OrganisationController(OrganisationService organisationService) {
         this.organisationService = organisationService;
     }
 
+    //TODO: rewrite comment
     /**
      * Returns a JSON object with all organisations in the table.
      *
@@ -42,6 +50,7 @@ public class OrganisationController {
         return organisationService.getAll();
     }
 
+    //TODO: rewrite comment
     /**
      * Returns a JSON object with description of the specified organization.
      *
@@ -50,9 +59,14 @@ public class OrganisationController {
      */
     @GetMapping("/organisations/{id}")
     public Organisation getOrganisation(@PathVariable("id") int organisationId) {
+        if (organisationService.get(organisationId) == null) {
+            throw new NoSuchEntityException("There is no organisation with ID = "
+                    + organisationId + " in the database.");
+        }
         return organisationService.get(organisationId);
     }
 
+    //TODO: rewrite comment
     /**
      * Accepts the JSON object with the description of the
      * organisation, transforms it and saves into the table.
@@ -68,6 +82,25 @@ public class OrganisationController {
     }
 
     /**
+     * Updates the existing organisation in the table. The request
+     * body must have got the JSON object with the ID of an existing
+     * entity and the updated data.
+     *
+     * @param organisation entity object with ID
+     * @return saved object
+     */
+    @PutMapping("/organisations")
+    public Organisation updateOrganisation(@RequestBody Organisation organisation) {
+        if (organisationService.get(organisation.getId()) == null) {
+            throw new NoSuchEntityException("There is no organisation with ID = "
+                    + organisation.getId() + " in the database.");
+        }
+        organisationService.saveOrUpdate(organisation);
+        return organisation;
+    }
+
+    //TODO: rewrite comment
+    /**
      * Deletes the organisation with the specified ID and return the
      * informative message.
      *
@@ -76,6 +109,10 @@ public class OrganisationController {
      */
     @DeleteMapping("/organisations/{id}")
     public String deleteUser(@PathVariable("id") int organisationId) {
+        if (organisationService.get(organisationId) == null) {
+            throw new NoSuchEntityException("There is no organisation with ID = "
+                    + organisationId + " in the database.");
+        }
         organisationService.delete(organisationId);
         return "Organisation with ID = " + organisationId + " was deleted.";
     }
